requirements.txt
pandas==1.3.1
Jinja2==2.11.2
openpyxl==3.0.7
python-dotenv==0.19.0

